# ICU-23061 LowercaseTransliterator has poor scalability due to synchronized state