--- a/icu4j/main/translit/src/main/java/com/ibm/icu/text/LowercaseTransliterator.java
+++ b/icu4j/main/translit/src/main/java/com/ibm/icu/text/LowercaseTransliterator.java
@@ -42,8 +42,6 @@ class LowercaseTransliterator extends Transliterator{
     private final ULocale locale;
 
     private final UCaseProps csp;
-    private ReplaceableContextIterator iter;
-    private StringBuilder result;
     private int caseLocale;
 
     /**
@@ -54,8 +52,6 @@ class LowercaseTransliterator extends Transliterator{
         super(_ID, null);
         locale = loc;
         csp=UCaseProps.INSTANCE;
-        iter=new ReplaceableContextIterator();
-        result = new StringBuilder();
         caseLocale = UCaseProps.getCaseLocale(locale);
     }
 
@@ -63,7 +59,7 @@ class LowercaseTransliterator extends Transliterator{
      * Implements {@link Transliterator#handleTransliterate}.
      */
     @Override
-    protected synchronized void handleTransliterate(Replaceable text,
+    protected void handleTransliterate(Replaceable text,
                                        Position offsets, boolean isIncremental) {
         if(csp==null) {
             return;
@@ -73,8 +69,10 @@ class LowercaseTransliterator extends Transliterator{
             return;
         }
 
+        ReplaceableContextIterator iter = new ReplaceableContextIterator();
+        StringBuilder result = new StringBuilder();
+
         iter.setText(text);
-        result.setLength(0);
         int c, delta;
 
         // Walk through original string