--- a/icu4j/perf-tests/src/main/java/com/ibm/icu/dev/test/perf/CaseFoldTransliteratorPerfTest.java
+++ b/icu4j/perf-tests/src/main/java/com/ibm/icu/dev/test/perf/CaseFoldTransliteratorPerfTest.java
@@ -1,0 +1,29 @@
+// © 2025 and later: Unicode, Inc. and others.
+// License & terms of use: http://www.unicode.org/copyright.html
+package com.ibm.icu.dev.test.perf;
+
+import java.util.concurrent.TimeUnit;
+
+import com.ibm.icu.text.Transliterator;
+import org.openjdk.jmh.annotations.Benchmark;
+import org.openjdk.jmh.annotations.BenchmarkMode;
+import org.openjdk.jmh.annotations.Mode;
+import org.openjdk.jmh.annotations.OutputTimeUnit;
+
+@BenchmarkMode(Mode.Throughput)
+@OutputTimeUnit(TimeUnit.MILLISECONDS)
+public class CaseFoldTransliteratorPerfTest {
+
+    static final Transliterator CASE = Transliterator.getInstance("CaseFold");
+
+    @Benchmark
+    public String testShort() {
+        return CASE.transliterate("Cat");
+    }
+
+    @Benchmark
+    public String testSentence() {
+        return CASE.transliterate("The Quick Brown Fox Jumped Over The Lazy Dog");
+    }
+
+}