@stevenschlansker 2025-02-27
I needed to add the license header to the performance test.