@richgillam 2025-03-18
The change itself looks reasonable, assuming the performance advantages are as billed.  I can't speak to the POM changes or the performance test, as that's not stuff I'm really familiar with.

Usually, bug fixes in ICU need to be done in both C++ and Java when appropriate.  Do we have a similar problem on the C++ side?  If so, I'd like to see a fix for that here too.