@grhoten 2025-03-21
These changes remind me of something. I think this Java lowercase transliterator still uses the locale.  The ICU4C implementation [removed that long ago](https://github.com/unicode-org/icu/commit/f6860c5a759a13482aa15700e082b7e647c698d8) because of how the registry works with sharing the singleton objects. Though the implementation in Java hard codes it to be US English casing only. So it's not much different.

I thought the transliterator casing functionality in ICU4C was doing something funny with thread safety long ago, but I can't seem to find the concerning code anymore. Perhaps it was removed, or I missed it.