@stevenschlansker 2025-03-21
> I can't speak to the POM changes or the performance test, as that's not stuff I'm really familiar with.

The introduction of JMH is optional, and if the project maintainers don't want to introduce it, I can remove these changes.
I see that ICU already has its own benchmark harness. JMH is nice because it handles a lot of the boilerplate for you - help messages, warming up the JVM, timing runs - in a way that the "JVM experts" have thought long and hard about. The benchmark could be converted to the ICU runner, or not committed, if that's preferred. But I think you'll find that more and more outside contributors would already come in with JMH familiarity.